+++ CHANGELOG.md
@@ -11,6 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ColorVariable` variant
 - `ColorVariableWatcher` component
 - `ColorReference` variant
+- `Vector2IntVariable` variant
+- `Vector2IntVariableWatcher` component
+- `Vector2IntReference` variant
+- `Vector2Variable` variant
+- `Vector2VariableWatcher` component
+- `Vector2Reference` variant
+- `Vector3Variable` variant
+- `Vector3VariableWatcher` component
+- `Vector3Reference` variant
+- `Vector4Variable` variant
+- `Vector4VariableWatcher` component
+- `Vector4Reference` variant
+- `SetPosition(Vector3Variable)` method to `TransformVariable` class
 - Custom editors for new types
 
 ### Changed